api-exercise/pyproject.toml
[tool.poetry]
name = "api-exercise"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
numpy = "^1.23.0"
torch = "^1.11.0"
torchvision = "^0.12.0"
timm = "^0.5.4"
Pillow = "^9.1.1"
pydantic = "^1.9.1"
pymongo = {extras = ["srv"], version = "^4.1.1"}
fastapi = "^0.78.0"
uvicorn = "^0.18.2"
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
